+++ NEWS.md
@@ -8,6 +8,7 @@
 * added the possibility to choose the shape of boxplot outlier points
 * added options for x and y scale tick formatting are now visible when pairs plot are active
 * added options for identity line and for placement of custom/target text at plot edges
+* added checkbox to recode numeic variables with -99 to missing
 
 # ggquickeda 0.2.1
 * added a new method for factor reordering min/max difference